Add FontAwesome modules and configuration Rename "types/" folder to "util/"
