--- src/types/configuration.ts
@@ -1,6 +1,0 @@
-/**
- * Global configuration.
- */
-export enum Config {
-  API_URL = 'http://localhost'
-}
